+++ lib/bound_mutation.dart
@@ -1,3 +1,4 @@
 library;
 
+export 'src/bound_action.dart' show BoundAction;
 export 'src/bound_mutation.dart' show BoundMutation;